SegregationSystem/requirements.txt
flask
requests
pylint
jsonschema
pytest
matplotlib
plotly
scikit-learn
kaleido
pillow